+++ Apps/Contoso.WinUI.Desktop.Puppet/Contoso.WinUI.Desktop.Puppet/MainWindow.xaml.cs
@@ -66,6 +66,11 @@ namespace Contoso.WinUI.Desktop.Puppet
             {
                 StorageMaxSize.Text = storageSize.ToString();
             }
+            var isManualSessionTrackerEnabled = localSettings.Values[Constants.KeyEnableManualSessionTracker] as bool?;
+            if (isManualSessionTrackerEnabled != null)
+            {
+                EnableManualSessionTrackerCheckBox.IsChecked = isManualSessionTrackerEnabled;
+            }
         }
 
         private void UpdateState()
@@ -439,5 +444,25 @@ namespace Contoso.WinUI.Desktop.Puppet
 
         [DllImport("user32.dll", ExactSpelling = true, CharSet = CharSet.Auto, PreserveSig = true, SetLastError = false)]
         public static extern IntPtr GetActiveWindow();
+
+        private void EnableManualSessionTrackerCheckBox_Checked(object sender, RoutedEventArgs e)
+        {
+            EnableManualSessionTracker(true);
+        }
+
+        private void EnableManualSessionTrackerCheckBox_Unchecked(object sender, RoutedEventArgs e)
+        {
+            EnableManualSessionTracker(false);
+        }
+
+        private void EnableManualSessionTracker(bool isEnabled)
+        {
+            localSettings.Values[Constants.KeyEnableManualSessionTracker] = isEnabled;
+        }
+
+        private void StartSessionButtonClick(object sender, RoutedEventArgs e)
+        {
+            Analytics.StartSession();
+        }
     }
 }