+++ SDK/AppCenterAnalytics/Microsoft.AppCenter.Analytics.Android/Analytics.cs
@@ -83,6 +83,22 @@ namespace Microsoft.AppCenter.Analytics
             AndroidAnalytics.TrackEvent(name, properties);
         }
 
+        /// <summary>
+        ///  Enable manual session tracker.
+        /// </summary>
+        public static void EnableManualSessionTracker()
+        {
+            AndroidAnalytics.EnableManualSessionTracker();
+        }
+
+        /// <summary>
+        /// Start a new session if manual session tracker is enabled, otherwise do nothing.
+        /// </summary>
+        public static void StartSession()
+        {
+            AndroidAnalytics.StartSession();
+        }
+
         ///// <summary>
         ///// Track a custom page.
         ///// </summary>