+++ SDK/AppCenterAnalytics/Microsoft.AppCenter.Analytics.MacOS/Analytics.cs
@@ -77,6 +77,22 @@ namespace Microsoft.AppCenter.Analytics
             MacOSAnalytics.TrackEvent(name);
         }
 
+        /// <summary>
+        ///  Enable manual session tracker.
+        /// </summary>
+        public static void EnableManualSessionTracker()
+        {
+            MacOSAnalytics.EnableManualSessionTracker();
+        }
+
+        /// <summary>
+        /// Start a new session if manual session tracker is enabled, otherwise do nothing.
+        /// </summary>
+        public static void StartSession()
+        {
+            MacOSAnalytics.StartSession();
+        }
+
         internal static void UnsetInstance()
         {
             MacOSAnalytics.ResetSharedInstance();